Move DeviceConfiguration to its own file
