+++ main/kernel/NetworkUtil.hpp
@@ -1,0 +1,29 @@
+#pragma once
+
+#include <string>
+
+#include <esp_mac.h>
+
+namespace farmhub::kernel {
+
+static const std::string& getMacAddress() {
+    static std::string macAddress;
+    if (macAddress.length() == 0) {
+        uint8_t rawMac[6];
+        for (int i = 0; i < 6; i++) {
+            rawMac[i] = 0;
+        }
+        if (esp_read_mac(rawMac, ESP_MAC_WIFI_STA) != ESP_OK) {
+            macAddress = "??:??:??:??:??:??:??:??";
+        } else {
+            char mac[24];
+            sprintf(mac, "%02x:%02x:%02x:%02x:%02x:%02x",
+                rawMac[0], rawMac[1], rawMac[2], rawMac[3],
+                rawMac[4], rawMac[5]);
+            macAddress = mac;
+        }
+    }
+    return macAddress;
+}
+
+}    // namespace farmhub::kernel
